feat(core): changed some utils functions' modifier from public to protected
