+++ packages/connector-alipay/src/index.test.ts
@@ -9,25 +9,30 @@ import {
   methodForAccessToken,
   methodForUserInfo,
 } from './constant';
-import {
-  mockedAlipayConfig,
-  mockedAlipayConfigWithValidPrivateKey,
-  mockedAlipayPublicParameters,
-  mockedTimestamp,
-} from './mock';
+import { mockedAlipayConfig, mockedAlipayPublicParameters, mockedTimestamp } from './mock';
 import { AlipayConfig } from './types';
+
+class AlipayConnectorTester extends AlipayConnector {
+  signingPamameters: jest.Mock = jest.fn();
+  getTimestamp: jest.Mock = jest.fn();
+
+  public getConfig: GetConnectorConfig<AlipayConfig>;
+
+  constructor(getConnectorConfig: GetConnectorConfig<AlipayConfig>) {
+    super(getConnectorConfig);
+
+    this.getConfig = getConnectorConfig;
+  }
+}
 
 const getConnectorConfig = jest.fn() as GetConnectorConfig<AlipayConfig>;
 
-const alipayMethods = new AlipayConnector(getConnectorConfig);
+const alipayMethods = new AlipayConnectorTester(getConnectorConfig);
 
 beforeAll(() => {
   jest.spyOn(alipayMethods, 'getConfig').mockResolvedValue(mockedAlipayConfig);
   jest.spyOn(alipayMethods, 'getTimestamp').mockReturnValue(mockedTimestamp);
 });
-
-const listenJSONParse = jest.spyOn(JSON, 'parse');
-const listenJSONStringify = jest.spyOn(JSON, 'stringify');
 
 const mockedSigningParameters = jest.spyOn(alipayMethods, 'signingPamameters');
 
@@ -48,58 +53,6 @@ describe('validateConfig', () => {
 
   it('should throw when missing required properties', async () => {
     await expect(alipayMethods.validateConfig({ appId: 'appId' })).rejects.toThrowError();
-  });
-});
-
-describe('signingParameters', () => {
-  afterEach(() => {
-    nock.cleanAll();
-    jest.clearAllMocks();
-  });
-
-  const testingParameters = {
-    ...mockedAlipayPublicParameters,
-    ...mockedAlipayConfigWithValidPrivateKey,
-    method: methodForAccessToken,
-    code: '7ffeb112fbb6495c9e7dfb720380DD39',
-  };
-
-  it('should return exact signature with the given parameters (functionality check)', () => {
-    const decamelizedParameters = alipayMethods.signingPamameters(testingParameters);
-    expect(decamelizedParameters.sign).toBe(
-      'td9+u0puul3HgbwLGL1X6z/vKKB/K25K5pjtLT/snQOp292RX3Y5j+FQUVuazTI2l65GpoSgA83LWNT9htQgtmdBmkCQ3bO6RWs38+2ZmBmH7MvpHx4ebUDhtebLUmHNuRFaNcpAZW92b0ZSuuJuahpLK8VNBgXljq+x0aD7WCRudPxc9fikR65NGxr5bwepl/9IqgMxwtajh1+PEJyhGGJhJxS1dCktGN0EiWXWNiogYT8NlFVCmw7epByKzCBWu4sPflU52gJMFHTdbav/0Tk/ZBs8RyP8Z8kcJA0jom2iT+dHqDpgkdzEmsR360UVNKCu5X7ltIiiObsAWmfluQ=='
-    );
-  });
-
-  it('should return exact signature with the given parameters (with empty property in testingParameters)', () => {
-    const decamelizedParameters = alipayMethods.signingPamameters({
-      ...testingParameters,
-      emptyProperty: '',
-    });
-    expect(decamelizedParameters.sign).toBe(
-      'td9+u0puul3HgbwLGL1X6z/vKKB/K25K5pjtLT/snQOp292RX3Y5j+FQUVuazTI2l65GpoSgA83LWNT9htQgtmdBmkCQ3bO6RWs38+2ZmBmH7MvpHx4ebUDhtebLUmHNuRFaNcpAZW92b0ZSuuJuahpLK8VNBgXljq+x0aD7WCRudPxc9fikR65NGxr5bwepl/9IqgMxwtajh1+PEJyhGGJhJxS1dCktGN0EiWXWNiogYT8NlFVCmw7epByKzCBWu4sPflU52gJMFHTdbav/0Tk/ZBs8RyP8Z8kcJA0jom2iT+dHqDpgkdzEmsR360UVNKCu5X7ltIiiObsAWmfluQ=='
-    );
-  });
-
-  it('should not call JSON.parse() when biz_content is empty', () => {
-    alipayMethods.signingPamameters(testingParameters);
-    expect(listenJSONParse).not.toHaveBeenCalled();
-  });
-
-  it('should call JSON.parse() when biz_content is not empty', () => {
-    alipayMethods.signingPamameters({
-      ...testingParameters,
-      biz_content: JSON.stringify({ AB: 'AB' }),
-    });
-    expect(listenJSONParse).toHaveBeenCalled();
-  });
-
-  it('should call JSON.stringify() when some value is object string', () => {
-    alipayMethods.signingPamameters({
-      ...testingParameters,
-      testObject: JSON.stringify({ AB: 'AB' }),
-    });
-    expect(listenJSONStringify).toHaveBeenCalled();
   });
 });
