+++ packages/connector-alipay/src/utils.ts
@@ -1,0 +1,51 @@
+import * as crypto from 'crypto';
+
+import * as iconv from 'iconv-lite';
+import snakeCaseKeys from 'snakecase-keys';
+
+import { alipaySigningAlgorithmMapping } from './constant';
+import { AlipayConfig } from './types';
+
+export type SigningPamameters = (
+  parameters: AlipayConfig & Record<string, string | undefined>
+) => Record<string, string>;
+
+// Reference: https://github.com/alipay/alipay-sdk-nodejs-all/blob/10d78e0adc7f310d5b07567ce7e4c13a3f6c768f/lib/util.ts
+export const signingPamameters: SigningPamameters = (
+  parameters: AlipayConfig & Record<string, string | undefined>
+): Record<string, string> => {
+  const { biz_content, privateKey, ...rest } = parameters;
+  const signParameters = snakeCaseKeys(
+    biz_content
+      ? {
+          ...rest,
+          bizContent: JSON.stringify(snakeCaseKeys(JSON.parse(biz_content))),
+        }
+      : rest
+  );
+
+  const decamelizeParameters = snakeCaseKeys(signParameters);
+
+  // eslint-disable-next-line @silverhand/fp/no-mutating-methods
+  const sortedParametersAsString = Object.entries(decamelizeParameters)
+    .map(([key, value]) => {
+      // Supported Encodings can be found at https://github.com/ashtuchkin/iconv-lite/wiki/Supported-Encodings
+
+      if (value) {
+        // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+        return `${key}=${iconv.encode(value, rest.charset ?? 'UTF8')}`;
+      }
+
+      return '';
+    })
+    .filter((keyValueString) => keyValueString)
+    .sort()
+    .join('&');
+
+  const sign = crypto
+    .createSign(alipaySigningAlgorithmMapping[rest.signType])
+    .update(sortedParametersAsString, 'utf8')
+    .sign(privateKey, 'base64');
+
+  return { ...decamelizeParameters, sign };
+};
